requirements.txt
numpy
pandas
python-dotenv
mysql-connector-python
pymysql
scikit-learn
dnspython==2.2.1
database-connect==0.1.66
pymongo==4.2.0
python-dotenv==0.21.0